Refactor S3/Lambda/Blob under src/endpoint and separate each op to its own module + bug fixes
s3: split controller ops

NOTE: this commit is not standalone as it contains only the new ops files,
but without the changes in s3_rest.

block_store: fix noisy log prints

s3 ops - now with s3_rest

s3 ops fixes

s3 more fixes

s3 more error handling

s3 fix copy part

s3_get_object - move code to s3 op from object_io for better error handling

tools: tcp_tunnel parse ssl port option correctly

eslint and jseautify set options

tests: add a local-config option to SKIP_EXTERNAL_TESTS for dev

useful when one doesn't want to run unit-tests against AWS

endpoint: blob/s3 refactor fixes

ep: move lambda code to src/endpoint/

endpoint: move s3 and lambda to src/endpoint/

endpoint: rename server.js to endpoint.js

endpoint: rename blob files

endpoint: fix require paths after renames and some refactors

whitespace for readability

endpoint: fixes for parse_op_name and hide containers with dot in their name

endpoint: refactor common code, lambda signature fix

s3: copy fix parsing to handle encoded URI

remove top level scripts. will add db scripts if needed

bucket server allowed_buckets update using $push/$pullAll for concurrency

fix #2967 - serialize system_store.load()

remove s3_controller after rebase

s3 validate client ip against account allowed ips

+++ src/test/unit_tests/test_system_servers.js
@@ -341,41 +341,43 @@ mocha.describe('system_servers', function() {
                     () => _.noop) // update bucket with 0 quota should fail
             )
             .then(() => {
+                if (config.SKIP_EXTERNAL_TESTS) return;
                 if (!process.env.AWS_ACCESS_KEY_ID ||
                     !process.env.AWS_SECRET_ACCESS_KEY) {
                     throw new Error('No valid AWS credentials in env - ' +
                         'AWS_ACCESS_KEY_ID + AWS_SECRET_ACCESS_KEY are required ' +
                         'for testing account.add_external_connection()');
                 }
+                return P.resolve()
+                    .then(() => client.account.add_external_connection({
+                        name: CLOUD_SYNC_CONNECTION,
+                        endpoint: 'https://s3.amazonaws.com',
+                        endpoint_type: 'AWS',
+                        identity: process.env.AWS_ACCESS_KEY_ID,
+                        secret: process.env.AWS_SECRET_ACCESS_KEY
+                    }))
+                    .then(() => client.account.delete_external_connection({
+                        connection_name: CLOUD_SYNC_CONNECTION,
+                    }))
+                    .then(() => client.account.add_external_connection({
+                        name: CLOUD_SYNC_CONNECTION,
+                        endpoint: 'https://s3.amazonaws.com',
+                        endpoint_type: 'AWS',
+                        identity: process.env.AWS_ACCESS_KEY_ID,
+                        secret: process.env.AWS_SECRET_ACCESS_KEY
+                    }))
+                    .then(() => client.bucket.set_cloud_sync({
+                        name: BUCKET,
+                        connection: CLOUD_SYNC_CONNECTION,
+                        target_bucket: BUCKET,
+                        policy: {
+                            schedule_min: 11
+                        }
+                    }));
+                    // .then(() => client.bucket.get_cloud_buckets({
+                    //     connection: CLOUD_SYNC_CONNECTION
+                    // }))
             })
-            .then(() => client.account.add_external_connection({
-                name: CLOUD_SYNC_CONNECTION,
-                endpoint: 'https://s3.amazonaws.com',
-                endpoint_type: 'AWS',
-                identity: process.env.AWS_ACCESS_KEY_ID,
-                secret: process.env.AWS_SECRET_ACCESS_KEY
-            }))
-            .then(() => client.account.delete_external_connection({
-                connection_name: CLOUD_SYNC_CONNECTION,
-            }))
-            .then(() => client.account.add_external_connection({
-                name: CLOUD_SYNC_CONNECTION,
-                endpoint: 'https://s3.amazonaws.com',
-                endpoint_type: 'AWS',
-                identity: process.env.AWS_ACCESS_KEY_ID,
-                secret: process.env.AWS_SECRET_ACCESS_KEY
-            }))
-            .then(() => client.bucket.set_cloud_sync({
-                name: BUCKET,
-                connection: CLOUD_SYNC_CONNECTION,
-                target_bucket: BUCKET,
-                policy: {
-                    schedule_min: 11
-                }
-            }))
-            // .then(() => client.bucket.get_cloud_buckets({
-            //     connection: CLOUD_SYNC_CONNECTION
-            // }))
             .then(() => client.system.read_system())
             // .then(() => client.bucket.get_cloud_sync({
             //     name: BUCKET,
